+++ adventcode/2024/04/BotD4b.cs
@@ -21,7 +21,7 @@ namespace adventcode._2024
         };
 
 
-        public override int Compute(IEnumerable<string> datas)
+        public override long Compute(IEnumerable<string> datas)
         {
             char[,] grid = ConvertToGrid(datas.ToArray());
